Реализовал CRUD категории, реализовал темную тему и перевел все на BootStrap

+++ ContactList/VebPractika/Pages/Contacts/Delete.cshtml.cs
@@ -1,5 +1,6 @@
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
+using Microsoft.EntityFrameworkCore;
 using ContactList.Data;
 using ContactList.Model;
 
@@ -17,24 +18,28 @@ namespace ContactList.Pages.Contacts
         [BindProperty]
         public ContactList.Model.Contact Contact { get; set; }
 
-        public IActionResult OnGet(int id)
+        public async Task<IActionResult> OnGetAsync(int? id)
         {
-            Contact = _context.Contacts.Find(id);
+            if (id == null) return NotFound();
 
-            if (Contact == null)
-                return NotFound();
+            Contact = await _context.Contacts
+                .Include(c => c.Category)
+                .FirstOrDefaultAsync(m => m.Id == id);
+
+            if (Contact == null) return NotFound();
 
             return Page();
         }
 
-        public IActionResult OnPost()
+        public async Task<IActionResult> OnPostAsync(int? id)
         {
-            var contact = _context.Contacts.Find(Contact.Id);
+            if (id == null) return NotFound();
 
+            var contact = await _context.Contacts.FindAsync(id);
             if (contact != null)
             {
                 _context.Contacts.Remove(contact);
-                _context.SaveChanges();
+                await _context.SaveChangesAsync();
             }
 
             return RedirectToPage("Index");
